move badges and recent poors to stats screen
